+++ LookupAnything/DataParser.cs
@@ -500,18 +500,23 @@ namespace Pathoschild.Stardew.LookupAnything
             recipes.AddRange(
                 from entry in metadata.MachineRecipes
                 let machine = new SObject(Vector2.Zero, entry.MachineID)
+
+                from recipe in entry.Recipes
+                from output in recipe.PossibleOutputs
+                from outputId in output.Ids
+
                 select new RecipeModel(
                     key: null,
                     type: RecipeType.MachineInput,
                     displayType: machine.DisplayName,
-                    ingredients: entry.Ingredients.Select(p => new RecipeIngredientModel(p)),
-                    item: ingredient => this.CreateRecipeItem(ingredient?.ParentSheetIndex, entry.Output),
+                    ingredients: recipe.Ingredients.Select(p => new RecipeIngredientModel(p)),
+                    item: ingredient => this.CreateRecipeItem(ingredient?.ParentSheetIndex, outputId),
                     mustBeLearned: false,
-                    exceptIngredients: entry.ExceptIngredients?.Select(p => new RecipeIngredientModel(p)),
-                    outputItemIndex: entry.Output,
-                    minOutput: entry.MinOutput,
-                    maxOutput: entry.MaxOutput,
-                    outputChance: entry.OutputChance,
+                    exceptIngredients: recipe.ExceptIngredients?.Select(p => new RecipeIngredientModel(p)),
+                    outputItemIndex: outputId,
+                    minOutput: output.MinOutput,
+                    maxOutput: output.MaxOutput,
+                    outputChance: output.OutputChance,
                     machineParentSheetIndex: entry.MachineID,
                     isForMachine: p => p is SObject obj && obj.GetItemType() == ItemType.BigCraftable && obj.ParentSheetIndex == entry.MachineID
                 )
@@ -525,12 +530,12 @@ namespace Pathoschild.Stardew.LookupAnything
                     key: null,
                     type: RecipeType.BuildingBlueprint,
                     displayType: building.displayName,
-                    ingredients: entry.Ingredients.Select(p => new RecipeIngredientModel(p.Key, p.Value)),
+                    ingredients: entry.Ingredients.Select(p => new RecipeIngredientModel(new[] { p.Key }, p.Value)),
                     item: ingredient => this.CreateRecipeItem(ingredient?.ParentSheetIndex, entry.Output),
                     mustBeLearned: false,
                     outputItemIndex: entry.Output,
                     minOutput: entry.OutputCount ?? 1,
-                    exceptIngredients: entry.ExceptIngredients?.Select(p => new RecipeIngredientModel(p, 1)),
+                    exceptIngredients: entry.ExceptIngredients?.Select(p => new RecipeIngredientModel(new[] { p }, 1)),
                     machineParentSheetIndex: null,
                     isForMachine: p => p is Building target && target.buildingType.Value == entry.BuildingKey
                 )