+++ LookupAnything/Framework/Data/MachineRecipeIngredientData.cs
@@ -5,8 +5,11 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Data
     /// <summary>Metadata for an ingredient in a machine recipe.</summary>
     internal class MachineRecipeIngredientData
     {
-        /// <summary>The unique item ID.</summary>
-        public int ID { get; set; }
+        /*********
+        ** Accessors
+        *********/
+        /// <summary>The unique item IDs.</summary>
+        public int[] PossibleIds { get; set; }
 
         /// <summary>The number required (or <c>null</c> for the default).</summary>
         public int? Count { get; set; }