+++ LookupAnything/Framework/Fields/RecipesForMachineField.cs
@@ -174,70 +174,85 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Fields
         {
             foreach (RecipeModel recipe in recipes)
             {
-                // get inputs
-                List<EntryItem> inputs = new List<EntryItem>();
-                foreach (var ingredient in recipe.Ingredients)
+                bool isKnown = !recipe.MustBeLearned || recipe.KnowsRecipe(Game1.player);
+
+                foreach (int[] inputIds in this.GetCartesianInputs(recipe))
                 {
-                    int count = ingredient.Count;
-
-                    // category
-                    if (ingredient.ID < 0)
+                    // get inputs
+                    var inputs = new EntryItem[inputIds.Length];
+                    for (int i = 0; i < inputIds.Length; i++)
                     {
-                        Item sampleInput = gameHelper.GetObjectsByCategory(ingredient.ID).FirstOrDefault();
-                        if (sampleInput == null)
-                            continue;
+                        int id = inputIds[i];
+                        var ingredient = recipe.Ingredients[i];
 
-                        string displayText = this.GetItemDisplayText(name: sampleInput.getCategoryName(), minCount: count, maxCount: count, chance: 100);
-                        inputs.Add(new EntryItem
+                        // category
+                        if (id < 0)
                         {
-                            Sprite = null,
-                            DisplayText = displayText,
-                            DisplayTextSize = Game1.smallFont.MeasureString(displayText)
-                        });
-                    }
+                            Item sampleInput = gameHelper.GetObjectsByCategory(id).FirstOrDefault();
+                            if (sampleInput == null)
+                                continue;
 
-                    // item
-                    else
-                    {
-                        Item input = gameHelper.GetObjectBySpriteIndex(ingredient.ID);
-                        if (input is SObject obj)
-                        {
-                            if (ingredient.PreservedParentSheetIndex != null)
-                                obj.preservedParentSheetIndex.Value = ingredient.PreservedParentSheetIndex.Value;
-                            if (ingredient.PreserveType != null)
-                                obj.preserve.Value = ingredient.PreserveType.Value;
+                            string displayText = this.GetItemDisplayText(name: sampleInput.getCategoryName(), minCount: ingredient.Count, maxCount: ingredient.Count, chance: 100);
+                            inputs[i] = new EntryItem
+                            {
+                                Sprite = null,
+                                DisplayText = displayText,
+                                DisplayTextSize = Game1.smallFont.MeasureString(displayText)
+                            };
                         }
 
-                        string displayText = this.GetItemDisplayText(name: input.DisplayName, minCount: count, maxCount: count, chance: 100);
-                        inputs.Add(new EntryItem
+                        // item
+                        else
                         {
-                            Sprite = gameHelper.GetSprite(input),
+                            Item input = gameHelper.GetObjectBySpriteIndex(id);
+                            if (input is SObject obj)
+                            {
+                                if (ingredient.PreservedParentSheetIndex != null)
+                                    obj.preservedParentSheetIndex.Value = ingredient.PreservedParentSheetIndex.Value;
+                                if (ingredient.PreserveType != null)
+                                    obj.preserve.Value = ingredient.PreserveType.Value;
+                            }
+
+                            string displayText = this.GetItemDisplayText(name: input.DisplayName, minCount: ingredient.Count, maxCount: ingredient.Count, chance: 100);
+                            inputs[i] = new EntryItem
+                            {
+                                Sprite = gameHelper.GetSprite(input),
+                                DisplayText = displayText,
+                                DisplayTextSize = Game1.smallFont.MeasureString(displayText)
+                            };
+                        }
+                    }
+
+                    // get output
+                    EntryItem output;
+                    {
+                        Item outputItem = recipe.CreateItem(null);
+                        string displayText = this.GetItemDisplayText(name: outputItem.DisplayName, minCount: recipe.MinOutput, maxCount: recipe.MaxOutput, chance: recipe.OutputChance);
+                        output = new EntryItem
+                        {
+                            Sprite = gameHelper.GetSprite(outputItem),
                             DisplayText = displayText,
                             DisplayTextSize = Game1.smallFont.MeasureString(displayText)
-                        });
+                        };
                     }
-                }
 
-                // get output
-                EntryItem output;
-                {
-                    Item outputItem = recipe.CreateItem(null);
-                    string displayText = this.GetItemDisplayText(name: outputItem.DisplayName, minCount: recipe.MinOutput, maxCount: recipe.MaxOutput, chance: recipe.OutputChance);
-                    output = new EntryItem
+                    // build recipe
+                    yield return new Entry
                     {
-                        Sprite = gameHelper.GetSprite(outputItem),
-                        DisplayText = displayText,
-                        DisplayTextSize = Game1.smallFont.MeasureString(displayText)
+                        IsKnown = isKnown,
+                        Inputs = inputs.ToArray(),
+                        Output = output
                     };
                 }
-
-                yield return new Entry
-                {
-                    IsKnown = !recipe.MustBeLearned || recipe.KnowsRecipe(Game1.player),
-                    Inputs = inputs.ToArray(),
-                    Output = output
-                };
             }
+        }
+
+        /// <summary>Get the cartesian product of the possible input ingredients for a recipe.</summary>
+        /// <param name="recipe">The recipe whose input sets to list.</param>
+        private IEnumerable<int[]> GetCartesianInputs(RecipeModel recipe)
+        {
+            int[][] sets = recipe.Ingredients.Select(p => p.PossibleIds.ToArray()).ToArray();
+            return this.GetCartesianProduct(sets);
         }
 
         /// <summary>Get the display text for an input or output item.</summary>
@@ -261,6 +276,33 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Fields
                 text += " (" + I18n.Generic_Percent(chance) + ")";
 
             return text;
+        }
+
+        /// <summary>Get the cartesian product of an arbitrary number of arrays.</summary>
+        /// <typeparam name="T">The array value type.</typeparam>
+        /// <param name="arrays">The arrays to combine.</param>
+        /// <remarks>Derived from <a href="https://stackoverflow.com/a/33106054/262123">code by Peter Almazov</a>.</remarks>
+        private IEnumerable<T[]> GetCartesianProduct<T>(T[][] arrays)
+        {
+            int[] lengths = arrays.Select(a => a.Length).ToArray();
+            int length = arrays.Length;
+            int[] inds = new int[length];
+
+            while (inds[0] != lengths[0])
+            {
+                T[] result = new T[length];
+                for (int i = 0; i != length; i++)
+                    result[i] = arrays[i][inds[i]];
+                yield return result;
+
+                int j = length - 1;
+                inds[j]++;
+                while (j > 0 && inds[j] == lengths[j])
+                {
+                    inds[j--] = 0;
+                    inds[j]++;
+                }
+            }
         }
     }
 }