+++ LookupAnything/Framework/Lookups/Characters/CharacterSubject.cs
@@ -17,7 +17,6 @@ using StardewValley.Locations;
 using StardewValley.Monsters;
 using StardewValley.Network;
 using StardewValley.Objects;
-using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Characters
 {
@@ -27,11 +26,14 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Characters
         /*********
         ** Fields
         *********/
-        /// <summary>The NPC type.s</summary>
+        /// <summary>The NPC type.</summary>
         private readonly SubjectType TargetType;
 
         /// <summary>The lookup target.</summary>
         private readonly NPC Target;
+
+        /// <summary>Provides subject entries.</summary>
+        private readonly ISubjectRegistry Codex;
 
         /// <summary>Simplifies access to private game code.</summary>
         private readonly IReflectionHelper Reflection;
@@ -41,6 +43,9 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Characters
 
         /// <summary>Whether to highlight item gift tastes which haven't been revealed in the NPC profile.</summary>
         private readonly bool HighlightUnrevealedGiftTastes;
+
+        /// <summary>Whether to look up the original entity when the game spawns a temporary copy.</summary>
+        private readonly bool EnableTargetRedirection;
 
         /// <summary>Whether the NPC is Gourmand in the Fern Islands farm cave.</summary>
         private readonly bool IsGourmand;
@@ -56,6 +61,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Characters
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
+        /// <param name="codex">Provides subject entries.</param>
         /// <param name="gameHelper">Provides utility methods for interacting with the game code.</param>
         /// <param name="npc">The lookup target.</param>
         /// <param name="type">The NPC type.</param>
@@ -63,13 +69,16 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Characters
         /// <param name="reflectionHelper">Simplifies access to private game code.</param>
         /// <param name="progressionMode">Whether to only show content once the player discovers it.</param>
         /// <param name="highlightUnrevealedGiftTastes">Whether to highlight item gift tastes which haven't been revealed in the NPC profile.</param>
+        /// <param name="enableTargetRedirection">Whether to look up the original entity when the game spawns a temporary copy.</param>
         /// <remarks>Reverse engineered from <see cref="NPC"/>.</remarks>
-        public CharacterSubject(GameHelper gameHelper, NPC npc, SubjectType type, Metadata metadata, IReflectionHelper reflectionHelper, bool progressionMode, bool highlightUnrevealedGiftTastes)
+        public CharacterSubject(ISubjectRegistry codex, GameHelper gameHelper, NPC npc, SubjectType type, Metadata metadata, IReflectionHelper reflectionHelper, bool progressionMode, bool highlightUnrevealedGiftTastes, bool enableTargetRedirection)
             : base(gameHelper)
         {
+            this.Codex = codex;
             this.Reflection = reflectionHelper;
             this.ProgressionMode = progressionMode;
             this.HighlightUnrevealedGiftTastes = highlightUnrevealedGiftTastes;
+            this.EnableTargetRedirection = enableTargetRedirection;
 
             // initialize
             this.Target = npc;
@@ -314,7 +323,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Characters
             {
                 this.Reflection.GetMethod(trashBear, "updateItemWanted").Invoke();
                 int itemWantedIndex = this.Reflection.GetField<int>(trashBear, "itemWantedIndex").GetValue();
-                yield return new ItemIconField(this.GameHelper, I18n.TrashBearOrGourmand_ItemWanted(), new SObject(itemWantedIndex, 1));
+                yield return new ItemIconField(this.GameHelper, I18n.TrashBearOrGourmand_ItemWanted(), this.GameHelper.GetObjectBySpriteIndex(itemWantedIndex), this.Codex);
             }
 
             // show progress
@@ -325,6 +334,11 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Characters
         /// <param name="npc">The NPC for which to show info.</param>
         private IEnumerable<ICustomField> GetDataForVillager(NPC npc)
         {
+            // special case: Abigail in the mines is a temporary instance with the name
+            // 'AbigailMine', so the info shown will be incorrect.
+            if (this.EnableTargetRedirection && npc.Name == "AbigailMine" && npc.currentLocation?.Name == "UndergroundMine20")
+                npc = Game1.getCharacterFromName("Abigail") ?? npc;
+
             // social fields (birthday, friendship, gifting, etc)
             if (this.GameHelper.IsSocialVillager(npc))
             {