+++ LookupAnything/Framework/Models/RecipeIngredientModel.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using Pathoschild.Stardew.LookupAnything.Framework.Data;
 using StardewValley;
 using SObject = StardewValley.Object;
@@ -10,8 +11,8 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Models
         /*********
         ** Accessors
         *********/
-        /// <summary>The unique item ID.</summary>
-        public int ID { get; }
+        /// <summary>The unique item IDs that can be used for this ingredient slot.</summary>
+        public ISet<int> PossibleIds { get; }
 
         /// <summary>The number required.</summary>
         public int Count { get; }
@@ -27,13 +28,13 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Models
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
-        /// <param name="id">The unique item ID.</param>
+        /// <param name="possibleIds">The unique item IDs that can be used for this ingredient slot.</param>
         /// <param name="count">The number required.</param>
         /// <param name="preserveType">The <see cref="StardewValley.Object.preserve"/> value to match (or <c>null</c> to ignore it).</param>
         /// <param name="preservedParentSheetIndex">The <see cref="StardewValley.Object.preservedParentSheetIndex"/> value to match (or <c>null</c> to ignore it).</param>
-        public RecipeIngredientModel(int id, int count, SObject.PreserveType? preserveType = null, int? preservedParentSheetIndex = null)
+        public RecipeIngredientModel(int[] possibleIds, int count, SObject.PreserveType? preserveType = null, int? preservedParentSheetIndex = null)
         {
-            this.ID = id;
+            this.PossibleIds = new HashSet<int>(possibleIds);
             this.Count = count;
             this.PreserveType = preserveType;
             this.PreservedParentSheetIndex = preservedParentSheetIndex;
@@ -43,7 +44,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Models
         /// <param name="ingredient">The ingredient to copy.</param>
         public RecipeIngredientModel(MachineRecipeIngredientData ingredient)
             : this(
-                id: ingredient.ID,
+                possibleIds: ingredient.PossibleIds,
                 count: ingredient.Count ?? 1,
                 preserveType: ingredient.PreserveType,
                 preservedParentSheetIndex: ingredient.PreservedParentSheetIndex
@@ -59,7 +60,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Models
                 return false;
 
             // item fields
-            if (item.ParentSheetIndex != this.ID && item.Category != this.ID)
+            if (!this.PossibleIds.Contains(item.ParentSheetIndex) && !this.PossibleIds.Contains(item.Category))
                 return false;
 
             // object fields